+++ src/main/java/com/vaadin/flow/component/datepicker/GeneratedVaadinDatePicker.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.datepicker;
+package com.vaadin.flow.component.datepicker;
 
 import com.vaadin.ui.Component;
 import com.vaadin.ui.common.ComponentSupplier;
