+++ src/test/java/com/vaadin/flow/component/datepicker/DatePickerValidationPage.java
@@ -13,15 +13,21 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.datepicker;
+package com.vaadin.flow.component.datepicker;
 
-import com.vaadin.flow.testutil.AbstractValidationTest;
-import com.vaadin.flow.testutil.TestPath;
+import com.vaadin.flow.component.datepicker.DatePicker;
+import com.vaadin.flow.testutil.ValidationTestView;
+import com.vaadin.router.Route;
+import com.vaadin.ui.common.HasValidation;
 
 /**
- * Integration tests for {@link DatePicker} validation.
+ * View for testing validation with {@link DatePicker}.
  */
-@TestPath("date-picker-validation")
-public class DatePickerValidationPageIT extends AbstractValidationTest {
+@Route("date-picker-validation")
+public class DatePickerValidationPage extends ValidationTestView {
 
+    @Override
+    protected HasValidation getValidationComponent() {
+        return new DatePicker();
+    }
 }
